Revert "Versión con test al 100% de cobertura del UserService, UserWebService y User"
