+++ src/main/java/ar/edu/unq/devapps/grupoJ202301/backenddevappsapt/BackendDevappsAptApplication.java
@@ -1,10 +1,11 @@
-package ar.edu.unq.devapps.grupoj202301.backenddevappsapt;
+package ar.edu.unq.devapps.grupoJ202301.backenddevappsapt;
+
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration;
 
-@SpringBootApplication(exclude = SecurityAutoConfiguration.class)
+@SpringBootApplication
 public class BackendDevappsAptApplication {
+
 	public static void main(String[] args) {
 		SpringApplication.run(BackendDevappsAptApplication.class, args);
 	}
