+++ src/main/java/ar/edu/unq/devapps/grupoJ202301/backenddevappsapt/service/UserService.java
@@ -1,0 +1,5 @@
+package ar.edu.unq.devapps.grupoJ202301.backenddevappsapt.service;
+
+public interface UserService {
+    public void registerUser();
+}
