--- src/main/java/ar/edu/unq/devapps/grupoj202301/backenddevappsapt/service/UserService.java
@@ -1,7 +1,0 @@
-package ar.edu.unq.devapps.grupoj202301.backenddevappsapt.service;
-import ar.edu.unq.devapps.grupoj202301.backenddevappsapt.validation.exception.UserException;
-import ar.edu.unq.devapps.grupoj202301.backenddevappsapt.model.User;
-
-public interface UserService {
-    String register(User user) throws UserException;
-}
