--- src/main/java/ar/edu/unq/devapps/grupoj202301/backenddevappsapt/validation/exception/BusinessException.java
@@ -1,7 +1,0 @@
-package ar.edu.unq.devapps.grupoj202301.backenddevappsapt.validation.exception;
-
-public class BusinessException extends RuntimeException {
-    public BusinessException(String message) {
-        super(message);
-    }
-}
