--- src/main/java/ar/edu/unq/devapps/grupoj202301/backenddevappsapt/validation/exception/UserException.java
@@ -1,8 +1,0 @@
-package ar.edu.unq.devapps.grupoj202301.backenddevappsapt.validation.exception;
-
-public class UserException extends RuntimeException {
-
-    public UserException(String message) {
-        super(message);
-    }
-}
