+++ src/main/resources/application.properties
@@ -1,23 +1,9 @@
-######################## DEV ########################
-# spring.datasource.url = jdbc:postgresql://localhost:5432/devapps
-# spring.datasource.username = devapps
-# spring.datasource.password = pass
-# spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.PostgreSQLDialect
-# spring.jpa.hibernate.ddl-auto = update
+spring.datasource.url = jdbc:postgresql://localhost:5432/devapps
+spring.datasource.username = devapps
+spring.datasource.password = pass
 
-######################## SWAGGER ########################
-springdoc.swagger-ui.path=/doc/swagger-ui.html
-springdoc.api-docs.enabled=true
-springdoc.swagger-ui.enabled=true
-springdoc.pathsToMatch=/user/**
+# The SQL dialect makes Hibernate generate better SQL for the chosen database
+spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.PostgreSQLDialect
 
-server.port=8090
-spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
+# Hibernate ddl auto (create, create-drop, validate, update)
 spring.jpa.hibernate.ddl-auto = update
-spring.datasource.url=jdbc:h2:mem:devapps
-spring.datasource.username=sa
-spring.datasource.password=password
-spring.datasource.driverClassName=org.h2.Driver
-spring.jpa.generate-ddl=true
-spring.jpa.show-sql=true
-spring.jpa.properties.hibernate.format_sql=tr
