--- src/main/resources/sonar-project.properties
@@ -1,1 +1,0 @@
-sonar.coverage.jacoco.xmlReportPaths=target/site/jacoco/jacoco.xml
