+++ src/test/java/ar/edu/unq/devapps/grupoJ202301/backenddevappsapt/BackendDevappsAptApplicationTests.java
@@ -1,0 +1,13 @@
+package ar.edu.unq.devapps.grupoJ202301.backenddevappsapt;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class BackendDevappsAptApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
