--- src/test/resources/application.properties
@@ -1,4 +1,0 @@
-spring.datasource.url=jdbc:h2:mem:testdb;DB_CLOSE_DELAY=-1;DB_CLOSE_ON_EXIT=FALSE;MODE=MYSQL
-spring.datasource.username=sa
-spring.datasource.password=password
-spring.h2.console.enabled=true
